Initial commit of the work I have done so far.
Basically, the core functionality of the game is in place. Can go to and from levels, solve those levels and it is relatively easy to add levels. Main bit of functionality still needed is save data, then the rest is making it look nicer, adding more levels (and making level creation simpler) and UX changes.