Object pooling refactoring and using shapes and colours.
The object pool now passes a callback for the method to call to return it to a pool instead of having the return method be public. Still want a way of doing something similar the other way so that only the object pool can set the owner of a pooled object). The spinners use shapes as well as colours to help with colourblindness.