Did some refactoring of the puzzle screen and game data.
Took the code for handling what needs to be done with spinners and put it into its own file. Also changed the lists of colors and shapes in game data to be a single list of a type of object that has both a color and sprite. This way at the least there will not be a different number of colors and shapes.