Added version 0.1 of the puzzle editor tool.
Still a number of things I want to do with the tool (namely being able to test the puzzle without having to get it into the game) and there is almost certainly some amount of cleaning up of the code and UI, but at this point the basics of the tool (loading an existing puzzle data file, editing/deleting/recolourng existing stars and saving new or existing puzzles changes is in. So I am committing what I have done and additional stuff will be done in separate branches.