Made improvements to the popup system.
The graphic for the popup has been switched to something I put together (probably could be better, but I think it is at least better than the default.
The popup object is now a prefab that is instantiated.
And finally, I managed to figure out the layout thing. Turns out I just needed to check a box on the vertical layout group, no additional code needed.
(There was a couple issues with my tests from prefabs changing, that has been fixed and the tests pass again).