Additional improvements to the level select button.
Still not entirely happy with how the button looks in the middle, but I am not sure how to further improve it right now, so going to move onto another task for now.

Also fixed a bug with the audio where it did not switch to the normal puzzle music after selecting next level.